code review. Replacing a double with a decimal for Money Fields

+++ SplitProject.Domain/Models/Expense.cs
@@ -6,7 +6,7 @@
 
         public DateTime ExpenseDate { get; set; } = DateTime.Now;
         public string? ExpenseTitle { get; set; } = "Unknown Expense";
-        public double ExpenseAmount { get; set; }
+        public decimal ExpenseAmount { get; set; }
         public int UserId { get; set; } //Внешний ключ для Юзеров (ByUser)
         public User User { get; set; } //Кто совершил трату
         public List<Benefiter> Benefiters { get; set; } //На кого делят
